message-queue-log-producer/requirements.txt
flask==3.1.0
pika==1.3.2
pytest==8.3.4
requests==2.32.3
tenacity==8.2.3
pyyaml==6.0.2
